remove instagram query
